requirements.txt
python-mpv~=0.5.2
asciimatics~=1.13.0
readchar~=3.0.5
